ci: run python tests in a separate database
