+++ tests/base_tests.py
@@ -18,7 +18,7 @@
 """Unit tests for Superset"""
 import imp
 import json
-from contextlib import contextmanager
+from tests.fixtures.utils import TEST_DATABSE_NAME, get_test_database
 from typing import Any, Dict, Union, List, Optional
 from unittest.mock import Mock, patch
 
@@ -27,7 +27,6 @@ import pytest
 from flask import Response
 from flask_appbuilder.security.sqla import models as ab_models
 from flask_testing import TestCase
-from sqlalchemy.ext.declarative.api import DeclarativeMeta
 from sqlalchemy.orm import Session
 from sqlalchemy.sql import func
 
@@ -339,12 +338,6 @@ class SupersetTestCase(TestCase):
             ):
                 security_manager.del_permission_role(public_role, perm)
 
-    def _get_database_by_name(self, database_name="main"):
-        if database_name == "examples":
-            return get_example_database()
-        else:
-            raise ValueError("Database doesn't exist")
-
     def run_sql(
         self,
         sql,
@@ -352,7 +345,6 @@ class SupersetTestCase(TestCase):
         user_name=None,
         raise_on_error=False,
         query_limit=None,
-        database_name="examples",
         sql_editor_id=None,
         select_as_cta=False,
         tmp_table_name=None,
@@ -363,7 +355,7 @@ class SupersetTestCase(TestCase):
         if user_name:
             self.logout()
             self.login(username=(user_name or "admin"))
-        dbid = self._get_database_by_name(database_name).id
+        dbid = get_test_database().id
         json_payload = {
             "database_id": dbid,
             "sql": sql,
@@ -437,17 +429,12 @@ class SupersetTestCase(TestCase):
             db.session.commit()
 
     def validate_sql(
-        self,
-        sql,
-        client_id=None,
-        user_name=None,
-        raise_on_error=False,
-        database_name="examples",
+        self, sql, client_id=None, user_name=None, raise_on_error=False,
     ):
         if user_name:
             self.logout()
             self.login(username=(user_name if user_name else "admin"))
-        dbid = self._get_database_by_name(database_name).id
+        dbid = get_test_database().id
         resp = self.get_json_resp(
             "/superset/validate_sql_json/",
             raise_on_error=False,
@@ -525,16 +512,3 @@ class SupersetTestCase(TestCase):
         else:
             mock_method.assert_called_once_with("error", func_name)
         return rv
-
-
-@contextmanager
-def db_insert_temp_object(obj: DeclarativeMeta):
-    """Insert a temporary object in database; delete when done."""
-    session = db.session
-    try:
-        session.add(obj)
-        session.commit()
-        yield obj
-    finally:
-        session.delete(obj)
-        session.commit()
